pyproject.toml
[tool.poetry]
name = "car_rent_backend"
version = "0.1.0"
description = ""
authors = ["Divaltor <kuklavudu78@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
djangorestframework = "^3.12.4"
envparse = "^0.2.0"
psycopg2 = "^2.8.6"
djangorestframework-simplejwt = "^4.6.0"
django-seed = "^0.2.2"
django-extensions = "^3.1.3"
loguru = "^0.5.3"
django-cors-headers = "^3.7.0"
python-dateutil = "^2.8.1"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
faker-vehicle = "^0.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
